% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_heap}
\alias{build_heap}
\title{build_heap: Part of DCEM package.}
\usage{
build_heap(data)
}
\arguments{
\item{data}{(NumericMatrix): The dataset provided by the user.}
}
\value{
A NumericMatrix with the max heap property.
}
\description{
Implements the creation of heap. Internally called by the \code{dcem_star_train}.
}
\author{
Parichit Sharma \email{parishar@iu.edu}, Hasan Kurban, Mehmet Dalkilic
}
