% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcem_em_utils.R
\name{expectation_mv}
\alias{expectation_mv}
\title{expectation_mv: Part of DCEM package.}
\usage{
expectation_mv(data, weights, meu, sigma, prior, num_clusters, tolerance)
}
\arguments{
\item{data}{(matrix): The input data.}

\item{weights}{(matrix): The probability weight matrix.}

\item{meu}{(matrix): The matrix of meu.}

\item{sigma}{(list): The list of sigma (co-variance matrices).}

\item{prior}{(vector): The vector of priors.}

\item{num_clusters}{(numeric): The number of clusters.}

\item{tolerance}{(numeric): The system epsilon value.}
}
\value{
Updated probability weight matrix.
}
\description{
Calculates the probabilistic weights for the multivariate data.
}
