% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcem_em_utils.R
\name{expectation_uv}
\alias{expectation_uv}
\title{expectation_uv: Part of DCEM package.}
\usage{
expectation_uv(data, weights, meu, sigma, prior, num_clusters, tolerance)
}
\arguments{
\item{data}{(matrix): The input data.}

\item{weights}{(matrix): The probability weight matrix.}

\item{meu}{(vector): The vector of meu.}

\item{sigma}{(vector): The vector of sigma (standard-deviations).}

\item{prior}{(vector): The vector of priors.}

\item{num_clusters}{(numeric): The number of clusters.}

\item{tolerance}{(numeric): The system epsilon value.}
}
\value{
Updated probability weight matrix.
}
\description{
Calculates the probabilistic weights for the univariate data.
}
\references{
Using data to build a better EM: EM* for big data.

Hasan Kurban, Mark Jenne, Mehmet M. Dalkilic
(2016) <https://doi.org/10.1007/s41060-017-0062-1>.
}
\author{
Parichit Sharma \email{parishar@iu.edu}, Hasan Kurban, Mark Jenne, Mehmet Dalkilic

This work is partially supported by NCI Grant 1R01CA213466-01.
}
