% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcem_init_mv.R
\name{get_priors}
\alias{get_priors}
\title{get_priors: Part of DCEM package.}
\usage{
get_priors(num_priors)
}
\arguments{
\item{num_priors}{(numeric): Number of priors one corresponding to each cluster.}
}
\value{
A vector of uniformly initialized prior values (numeric).
}
\description{
Initialize the priors.
}
\details{
For example, if the user specify 2 priors then the vector will have 2
entries (one for each cluster) where each will be 1/2 or 0.5.
}
