% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcem_heap_util.R
\name{insert_nodes}
\alias{insert_nodes}
\title{insert_nodes: Part of DCEM package.}
\usage{
insert_nodes(heap_list, heap_assn, data_probs, leaves_ind, num_clusters)
}
\arguments{
\item{heap_list}{(list): The nested list containing the heaps. Each entry in the
list is a list maintained in max-heap structure.}

\item{heap_assn}{(numeric): The vector representing the heap assignments.}

\item{data_probs}{(string): A vector containing the probability for data.}

\item{leaves_ind}{(numeric): A vector containing the indices of leaves in heap.}

\item{num_clusters}{(numeric): The number of clusters. Default: \strong{2}}
}
\value{
A nested list. Each entry in the list is a list maintained
        in the max-heap structure.
}
\description{
Implements the node insertion into the heaps.
}
\references{
Parichit Sharma, Hasan Kurban, Mehmet Dalkilic DCEM: An R package for clustering big data via
data-centric modification of Expectation Maximization, SoftwareX, 17, 100944 URL
https://doi.org/10.1016/j.softx.2021.100944
}
