% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionosphere_data.R
\docType{data}
\name{ionosphere_data}
\alias{ionosphere_data}
\title{Ionosphere data: A dataset of 351 radar readings}
\format{
A file with 351 rows and 35 columns of multivariate data in a csv file. All values are numeric.
}
\source{
Space Physics Group
Applied Physics Laboratory
Johns Hopkins University
Johns Hopkins Road
Laurel, MD 20723
Web URL: \url{http://archive.ics.uci.edu/ml/datasets/Ionosphere}

 \strong{References:}
 Sigillito, V. G., Wing, S. P., Hutton, L. V., & Baker, K. B. (1989).
 Classification of radar returns from the ionosphere using neural networks.
 Johns Hopkins APL Technical Digest, 10, 262-266.
}
\usage{
ionosphere_data
}
\description{
This dataset contains 351 entries (radar readings from a system in goose bay laboratory) and 35 columns.
The 35th columns is the label columns identifying the entry as either good or bad. Additionally, the 2nd column
only contains 0's.
}
\keyword{datasets}
