% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_heapify}
\alias{max_heapify}
\title{max_heapify: Part of DCEM package.}
\usage{
max_heapify(data, index, num_data)
}
\arguments{
\item{data}{(NumericMatrix): The dataset provided by the user.}

\item{index}{(int): The index of the data point.}

\item{num_data}{(numeric): The total number of observations in the data.}
}
\value{
A NumericMatrix with the max heap property.
}
\description{
Implements the creation of max heap. Internally called by the \code{dcem_star_train}.
}
\author{
Parichit Sharma \email{parishar@iu.edu}, Hasan Kurban, Mehmet Dalkilic
}
