% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcem_em_utils.R
\name{maximisation_mv}
\alias{maximisation_mv}
\title{maximisation_mv: Part of DCEM package.}
\usage{
maximisation_mv(data, weights, meu, sigma, prior, num_clusters, num_data)
}
\arguments{
\item{data}{(matrix): The input data.}

\item{weights}{(matrix): The probability weight matrix.}

\item{meu}{(matrix): The matrix of meu.}

\item{sigma}{(list): The list of sigma (co-variance matrices).}

\item{prior}{(vector): The vector of priors.}

\item{num_clusters}{(numeric): The number of clusters.}

\item{num_data}{(numeric): The total number of observations in the data.}
}
\value{
Updated values for meu, sigma and prior.
}
\description{
Calculates meu, sigma and prior based on the updated probability weight matrix.
}
\references{
Using data to build a better EM: EM* for big data.

Hasan Kurban, Mark Jenne, Mehmet M. Dalkilic
(2016) <https://doi.org/10.1007/s41060-017-0062-1>.
}
\author{
Parichit Sharma \email{parishar@iu.edu}, Hasan Kurban, Mark Jenne, Mehmet Dalkilic

This work is partially supported by NCI Grant 1R01CA213466-01.
}
