% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcem_em_utils.R
\name{maximisation_uv}
\alias{maximisation_uv}
\title{maximisation_uv: Part of DCEM package.}
\usage{
maximisation_uv(data, weights, meu, sigma, prior, num_clusters, num_data)
}
\arguments{
\item{data}{(matrix): The input data.}

\item{weights}{(matrix): The probability weight matrix.}

\item{meu}{(vector): The vector of meu.}

\item{sigma}{(vector): The vector of sigma (standard-deviations).}

\item{prior}{(vector): The vector of priors.}

\item{num_clusters}{(numeric): The number of clusters.}

\item{num_data}{(numeric): The total number of observations in the data.}
}
\value{
Updated values for meu, sigma and prior.
}
\description{
Calculates meu, sigma and prior based on the updated probability weight matrix.
}
