% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcem_init_mv.R
\name{meu_mv}
\alias{meu_mv}
\title{meu_mv: Part of DCEM package.}
\usage{
# Randomly seeding the mean(s).
meu_mv(data, num_meu)
}
\arguments{
\item{data}{(matrix): The dataset provided by the user.}

\item{num_meu}{(numeric): The number of meu.}
}
\value{
A matrix containing the selected samples from the dataset.
}
\description{
Initialize the meus(s) by randomly selecting the samples from the dataset. This is the
\strong{default} method for initializing the meu(s).
}
