% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcem_init_mv.R
\name{meu_mv_impr}
\alias{meu_mv_impr}
\title{meu_mv_impr: Part of DCEM package.}
\usage{
# Randomly seeding the meu.
meu_mv_impr(data, num_meu)
}
\arguments{
\item{data}{(matrix): The dataset provided by the user.}

\item{num_meu}{(numeric): The number of meu.}
}
\value{
A matrix containing the selected samples from the dataset.
}
\description{
Initialize the meu(s) by randomly selecting the samples from the dataset. It uses the proposed
implementation from K-means++: The Advantages of Careful Seeding, David Arthur and Sergei
Vassilvitskii. URL http://ilpubs.stanford.edu:8090/778/1/2006-13.pdf.
}
