% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcem_init_uv.R
\name{meu_uv}
\alias{meu_uv}
\title{meu_uv: Part of DCEM package.}
\usage{
# Randomly seeding the meu.
meu_uv(data, num_meu)
}
\arguments{
\item{data}{(matrix): The dataset provided by the user.}

\item{num_meu}{(number): The number of meu.}
}
\value{
A vector containing the selected samples from the dataset.
}
\description{
This function is internally called by the dcem_train to initialize the
meu(s). It randomly selects the meu(s) from the
range min(data):max(data).
}
