% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcem_init_uv.R
\name{meu_uv_impr}
\alias{meu_uv_impr}
\title{meu_uv_impr: Part of DCEM package.}
\usage{
# Seeding the meu using the K-means++ implementation.
meu_uv_impr(data, num_meu)
}
\arguments{
\item{data}{(matrix): The dataset provided by the user.}

\item{num_meu}{(number): The number of meu.}
}
\value{
A vector containing the selected samples from the dataset.
}
\description{
This function is internally called by the dcem_train to initialize the
meu(s). It uses the proposed implementation from
K-means++: The Advantages of Careful Seeding, David Arthur and Sergei Vassilvitskii.
URL http://ilpubs.stanford.edu:8090/778/1/2006-13.pdf.
}
