% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcem_init_mv.R
\name{priors}
\alias{priors}
\title{priors: Part of DCEM package.}
\usage{
priors(num)
}
\arguments{
\item{num}{(numeric): Number of priors.}
}
\value{
A vector of uniformly initialized prior values (numeric).
}
\description{
Initializes the prior vectors for the Gaussian(s)
}
\details{
For example, if the user specify 2 priors then the vector will have 2
entries (one for each cluster) where each will be 1/2 or 0.5.
}
\examples{
#Randomly generating the priors.

priors(2)
priors(3)

}
\author{
Parichit Sharma \email{parishar@iu.edu}, Hasan Kurban, Mark Jenne, Mehmet Dalkilic

This work was partially supported by NCI Grant 1R01CA213466-01.
}
