% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcem_heap_util.R
\name{separate_data}
\alias{separate_data}
\title{separate_data: Part of DCEM package.}
\usage{
separate_data(heap_list, num_clusters)
}
\arguments{
\item{heap_list}{(list): The nested list containing the heaps. Each entry in the
list is a list maintained in max-heap structure.}

\item{num_clusters}{(numeric): The number of clusters. Default: \strong{2}}
}
\value{
A nested list where,

        First entry is the list of heaps with leaves removed.

        Second entry is the list of leaves.
}
\description{
Separate leaf nodes from the heaps.
}
\references{
Parichit Sharma, Hasan Kurban, Mehmet Dalkilic DCEM: An R package for clustering big data via
data-centric modification of Expectation Maximization, SoftwareX, 17, 100944 URL
https://doi.org/10.1016/j.softx.2021.100944
}
