% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcem_init_mv.R
\name{sigma_mv}
\alias{sigma_mv}
\title{sigma_mv: Part of DCEM package.}
\usage{
sigma_mv(num_sigma, numcol)
}
\arguments{
\item{num_sigma}{(numeric): Number of covariance matrices.}

\item{numcol}{(numeric): The number of columns in the dataset.}
}
\value{
A list of identity matrices. The number of entries in the list
        is equal to the input parameter (num_cov).
}
\description{
Initializes the co-variance matrices as the identity matrices.
}
