% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcem_init_uv.R
\name{sigma_uv}
\alias{sigma_uv}
\title{sigma_uv: Part of DCEM package.}
\usage{
sigma_uv(data, num_sigma)
}
\arguments{
\item{data}{(matrix): The dataset provided by the user.}

\item{num_sigma}{(number): Number of sigma (standard_deviations).}
}
\value{
A vector of standard deviation value(s).
}
\description{
Initializes the standard deviation for the Gaussian(s).
}
