% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitycheck.R
\name{trim_data}
\alias{trim_data}
\title{trim_data: Part of DCEM package. Used internally in the package.}
\usage{
trim_data(columns, data)
}
\arguments{
\item{columns}{(string): A comma separated
list of column(s) that needs to be removed from the dataset.
Default: ''}

\item{data}{(dataframe): Dataframe containing the input data.}
}
\value{
A dataframe with the specified column(s) removed from it.
}
\description{
Removes the specified column(s) from the dataset.
}
