% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcem_em_utils.R
\name{update_weights}
\alias{update_weights}
\title{update_weights: Part of DCEM package.}
\usage{
update_weights(temp_weights, weights, index_list, num_clusters)
}
\arguments{
\item{temp_weights}{(matrix): A matrix of probabilistic weights for leaf data.}

\item{weights}{(matrix): A matrix of probabilistic weights for all data.}

\item{index_list}{(vector): A vector of indices.}

\item{num_clusters}{(numeric): The number of clusters.}
}
\value{
Updated probabilistic weights matrix.
}
\description{
Update the probability values for specific data points that change between the heaps.
}
