% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcem_heap_util.R
\name{insert_nodes}
\alias{insert_nodes}
\title{insert_nodes: Part of DCEM package.}
\usage{
insert_nodes(heap_list, heap_assn, data_probs, leaves_ind, num_clusters)
}
\arguments{
\item{heap_list}{(list): The nested list containing the heaps. Each entry in the
list is a list maintained in max-heap structure.}

\item{heap_assn}{(numeric): The vector representing the heap assignments.}

\item{data_probs}{(string): A vector containing the probability for data.}

\item{leaves_ind}{(numeric): A vector containing the indices of leaves in heap.}

\item{num_clusters}{(numeric): The number of clusters. Default: \strong{2}}
}
\value{
A nested list. Each entry in the list is a list maintained
        in the max-heap structure.
}
\description{
Implements the node insertion into the heaps.
}
\references{
Using data to build a better EM: EM* for big data.

Hasan Kurban, Mark Jenne, Mehmet M. Dalkilic
(2016) <https://doi.org/10.1007/s41060-017-0062-1>.
}
\author{
Parichit Sharma \email{parishar@iu.edu}, Hasan Kurban, Mark Jenne, Mehmet Dalkilic

This work is partially supported by NCI Grant 1R01CA213466-01.
}
