% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcem_heap_util.R
\name{separate_data}
\alias{separate_data}
\title{separate_data: Part of DCEM package.}
\usage{
separate_data(heap_list, num_clusters)
}
\arguments{
\item{heap_list}{(list): The nested list containing the heaps. Each entry in the
list is a list maintained in max-heap structure.}

\item{num_clusters}{(numeric): The number of clusters. Default: \strong{2}}
}
\value{
A nested list where,

        First entry is the list of heaps with leaves removed.

        Second entry is the list of leaves.
}
\description{
Separate leaf nodes from the heaps.
}
\references{
Using data to build a better EM: EM* for big data.

Hasan Kurban, Mark Jenne, Mehmet M. Dalkilic
(2016) <https://doi.org/10.1007/s41060-017-0062-1>.
}
\author{
Parichit Sharma \email{parishar@iu.edu}, Hasan Kurban, Mark Jenne, Mehmet Dalkilic

This work is partially supported by NCI Grant 1R01CA213466-01.
}
