% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitycheck.R
\name{trim_data}
\alias{trim_data}
\title{trim_data: Part of DCEM package.}
\usage{
trim_data(columns, data)
}
\arguments{
\item{columns}{(string): A comma separated
list of column(s) that needs to be removed from the dataset.
Default: ''}

\item{data}{(dataframe): Dataframe containing the input data.}
}
\value{
A dataframe with the specified column(s) removed from it.
}
\description{
Removes the specified column(s) from the dataset.
}
\examples{
# Remove a range of columns. Generally, the columns containing the labels or
# redundant values (such as all 0's) should be removed before training the model.

trim_data("1,2", data.frame(x1=sample(1:100,10),
x2=sample(500:1000, 10), x3=sample(-100:0,10)))

# Remove a single column.

trim_data("2", data.frame(x1=sample(1:100,10),
x2=sample(500:1000, 10), x3=sample(-100:0,10)))

}
\references{
Using data to build a better EM: EM* for big data.

Hasan Kurban, Mark Jenne, Mehmet M. Dalkilic
(2016) <doi:https://doi.org/10.1007/s41060-017-0062-1>.
}
\author{
Parichit Sharma \email{parishar@iu.edu}, Hasan Kurban, Mark Jenne, Mehmet Dalkilic
This work is partially supported by NCI Grant 1R01CA213466-01.
}
