% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcem_em_utils.R
\name{update_weights}
\alias{update_weights}
\title{update_weights: Part of DCEM package.}
\usage{
update_weights(temp_weights, weights, index_list, num_clusters)
}
\arguments{
\item{temp_weights}{(matrix): A matrix of probabilistic weights for leaf data.}

\item{weights}{(matrix): A matrix of probabilistic weights for all data.}

\item{index_list}{(vector): A vector of indices.}

\item{num_clusters}{(numeric): The number of clusters.}
}
\value{
Updated probabilistic weights matrix.
}
\description{
Update the probability values for specific data points that change between the heaps.
}
\references{
Using data to build a better EM: EM* for big data.

Hasan Kurban, Mark Jenne, Mehmet M. Dalkilic
(2016) <https://doi.org/10.1007/s41060-017-0062-1>.
}
\author{
Parichit Sharma \email{parishar@iu.edu}, Hasan Kurban, Mark Jenne, Mehmet Dalkilic

This work is partially supported by NCI Grant 1R01CA213466-01.
}
