% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitycheck.R
\name{validate_data}
\alias{validate_data}
\title{validate_data: Part of DCEM package. Used internally in the package.}
\usage{
validate_data(columns, numcols)
}
\arguments{
\item{columns}{(string): A comma separated
list of columns that needs to be removed from the dataset. Default: ''}

\item{numcols}{(numeric): Number of columns in the dataset.}
}
\value{
boolean: TRUE if the columns exists otherwise FALSE.
}
\description{
Implements sanity check for the input data. This function is for internal use and is called
by the \code{\link{dcem_train}}.
}
\details{
An example would be to check if the column to be removed exist
or not? \code{\link{trim_data}} internally calls this function before removing
the column(s).
}
